+++ src/components/MapView.jsx
@@ -12,7 +12,6 @@ import {imageURL} from '../utils/utils';
 // MUI
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
-import Link from '@mui/material/Link';
 import Button from '@mui/material/Button';
 import Chip from '@mui/material/Chip';
 import IconButton from '@mui/material/IconButton';