+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.10.0] - 2023-06-25
+
+Add:
+  - Admin: New centralised Listing component for all content types
+  - Admin: Filterable content tables
+  - Admin: Restruct admin components
+
 ## [1.9.0] - 2023-06-23
 
 Add: